+++ include/hpp/fcl/fwd.hh
@@ -70,6 +70,19 @@
     throw exception(ss.str());                                \
   }
 
+#ifdef HPP_FCL_TURN_ASSERT_INTO_EXCEPTION
+#define HPP_FCL_ASSERT(check, message, exception)         \
+  {                                                       \
+    if (!check) HPP_FCL_THROW_PRETTY(message, exception); \
+  }
+#else
+#define HPP_FCL_ASSERT(check, message, exception) \
+  {                                               \
+    HPP_FCL_UNUSED_VARIABLE(exception(message));  \
+    assert(check&& message);                      \
+  }
+#endif
+
 #if (__cplusplus >= 201103L || (defined(_MSC_VER) && _MSC_VER >= 1600))
 #define HPP_FCL_WITH_CXX11_SUPPORT
 #endif